Split the review stepper into focused modules
The stepper component had grown to ~600 lines. Extract the parts that stand on
their own, leaving the component as the view (~395 lines):

- markdown-review-stepper-state.ts: reducer, state types, and the decision
  carry-over helpers (rehydrate/record).
- markdown-review-stepper-dom.ts: active-block highlight/reveal and the platform
  shortcut hint.
- markdown-review-stepper-keyboard.ts: the scoped keyboard-shortcuts hook.

No behavior change; covered by the existing stepper unit tests and E2E.
